+++ gribmagic/commands.py
@@ -9,6 +9,7 @@ import click
 
 import gribmagic.dwd.download
 import gribmagic.installer
+import gribmagic.smith.bbox
 from gribmagic.unity.configuration.parser import parse_model_config
 from gribmagic.unity.core import run_model_download
 from gribmagic.unity.enumerations import WeatherModel
@@ -29,20 +30,20 @@ def cli(ctx):
     pass
 
 
-@cli.group(help="Unified NWP data downloader")
+@cli.group(help="Unified NWP data downloader.")
 @click.pass_context
 def unity(ctx):
     pass
 
 
-@unity.command(name="list", help="List available NWP data labels")
+@unity.command(name="list", help="List available NWP data labels.")
 def unity_list():
     model_config = parse_model_config()
     labels = [label for label in model_config.keys() if not label.endswith("-base")]
     print(json.dumps(labels, indent=4))
 
 
-@unity.command(name="acquire", help="Acquire NWP data")
+@unity.command(name="acquire", help="Acquire NWP data.")
 @click.option("--model", required=True, help="The weather model name.")
 @click.option("--timestamp", required=True, help="The initialization timestamp.")
 @click.option("--target", required=True, envvar="GM_DATA_PATH", help="The target directory.")
@@ -62,7 +63,7 @@ def unity_acquire(
     list(results)
 
 
-@cli.group(help="DWD GRIB data downloader")
+@cli.group(help="DWD GRIB data downloader.")
 @click.pass_context
 def dwd(ctx):
     pass
@@ -70,3 +71,12 @@ def dwd(ctx):
 
 dwd.add_command(gribmagic.dwd.download.main, "acquire")
 cli.add_command(gribmagic.installer.main, "install")
+
+
+@cli.group(help="Toolsmith.")
+@click.pass_context
+def smith(ctx):
+    pass
+
+
+smith.add_command(gribmagic.smith.bbox.main, "bbox")